single product dark mode added, router dom goes one stage forward.
