+++ countries-app/src/components/BackButton/BackButton.jsx
@@ -1,8 +1,10 @@
 import { BiArrowBack } from "react-icons/bi";
-function BackButton() {
+function BackButton({ darkMode }) {
   return (
     <div
-      className="d-flex justify-content-center align-items-center py-2 gap-3 mb-5 shadow rounded bg-white"
+      className={`d-flex justify-content-center align-items-center py-2 gap-3 mb-5 shadow rounded ${
+        darkMode ? "dark-cards" : "light-cards"
+      }`}
       style={{ width: "150px" }}
     >
       <BiArrowBack style={{ fontSize: "1.2rem" }} />
