+++ countries-app/src/index.js
@@ -1,18 +1,14 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import { BrowserRouter, Route, Routes } from "react-router-dom";
-import "./index.css";
 import App from "./App";
-import SingleProduct from "./components/SingleProduct/SingleProduct";
+import "./index.css";
+import { BrowserRouter } from "react-router-dom";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
   <React.StrictMode>
     <BrowserRouter>
-      <Routes>
-        <Route exact path="/" element={<App />} />
-        <Route exact path="/name/:countryName" element={<SingleProduct />} />
-      </Routes>
+      <App />
     </BrowserRouter>
   </React.StrictMode>
 );
